Add files via upload
the deleteAllWords button works, the custom definition thing is stored, the words with their definitions gets displayed, and I added a better looking font. I uploaded all the working code to the main branch
Only thing left to do is to try to implement the portion where if this word is in the database, it's highlighted on other pages.